Refactor email service, remove html code from java file, and changed access_token expiration time to 1 hour
