+++ Backend/microservices/authority/pom.xml
@@ -79,5 +79,11 @@
             <artifactId>spring-boot-starter-validation</artifactId>
             <version>2.6.6</version>
         </dependency>
+
+        <dependency>
+            <groupId>org.thymeleaf</groupId>
+            <artifactId>thymeleaf-spring5</artifactId>
+            <version>3.0.15.RELEASE</version>
+        </dependency>
     </dependencies>
 </project>
