--- Backend/microservices/authority/src/main/java/com/ms/authority/email/EmailService.java
@@ -1,16 +1,0 @@
-package com.ms.authority.email;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.scheduling.annotation.Async;
-
-import javax.mail.MessagingException;
-
-public interface EmailService {
-    void sendSimpleMessage(String to, String subject, String text);
-
-    void sendMessageWithAttachment(
-            String to, String subject, String text, String pathToAttachment) throws MessagingException;
-
-    @Async
-    void sendActivationLink(String to, String email, String link) throws MessagingException;
-}
